backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
websockets
mcp
google-generativeai
bcrypt
python-multipart
pydantic
pydantic-settings
openai